•	Cruise Database Management System (Sept. 2011 – Dec. 2011) and Canteen Database Management System (Sept. 2010 - Jan. 2011)
Institution : HKBK College of Engineering, Bangalore
Abstract : The objective and scope of the Project was to collect all the
related information about a customer’s voyage. The software involves all
the facilities provided at the customers’ disposal. The software is used
to record the customer’s personal details and expenses, and to compute
the bills